+++ src/api/endpoints.ts
@@ -15,6 +15,8 @@ export const ENDPOINTS = {
         TWO_FACTOR_VERIFY: `${API_VERSION_PREFIX}/auth/two-factor-verify`,
         LOGOUT: `${API_VERSION_PREFIX}/auth/logout`,
         REFRESH: `${API_VERSION_PREFIX}/auth/refresh-token`,
+        FORGOT_PASSWORD: `${API_VERSION_PREFIX}/auth/forgot-password`,
+        RESET_PASSWORD: `${API_VERSION_PREFIX}/auth/reset-password`,
         SET_LANGUAGE: `${API_VERSION_PREFIX}/auth/set-language`,
         USER_DATA: `${API_VERSION_PREFIX}/auth/user-data`,
         EVENTS: `${API_VERSION_PREFIX}/auth/events`,